# sei-sidecar/api/openapi.yaml
openapi: 3.1.0
info:
  title: sei-sidecar API
  description: HTTP API for the sei-sidecar task executor and scheduler.
  version: 0.1.0
  license:
    name: Apache-2.0

servers:
  - url: http://localhost:7777

paths:
  /healthz:
    get:
      operationId: healthz
      summary: Liveness probe
      description: Returns 200 after mark-ready has completed; 503 otherwise.
      responses:
        "200":
          description: Sidecar is ready.
        "503":
          description: Sidecar is not yet ready.

  /status:
    get:
      operationId: getStatus
      summary: Engine status snapshot
      responses:
        "200":
          description: Current engine state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineStatus"

  /task:
    post:
      operationId: submitTask
      summary: Submit a task for execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "202":
          description: Task accepted.
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [running]
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Resource conflict — task cannot be scheduled right now.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks:
    get:
      operationId: listTasks
      summary: List tracked tasks
      parameters:
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/TaskState"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        "200":
          description: Task history (newest first).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackedTask"

  /tasks/{taskId}:
    get:
      operationId: getTask
      summary: Get a single tracked task by ID
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedTask"
        "404":
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /schedules:
    post:
      operationId: createSchedule
      summary: Create a time-based schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleRequest"
      responses:
        "201":
          description: Schedule created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Invalid schedule definition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listSchedules
      summary: List all schedules
      responses:
        "200":
          description: All schedules.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"

  /schedules/{scheduleId}:
    get:
      operationId: getSchedule
      summary: Get a single schedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Schedule details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Schedule not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSchedule
      summary: Remove a schedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Schedule deleted.
        "404":
          description: Schedule not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    TaskRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          description: Task type identifier.
        params:
          type: object
          additionalProperties: true

    TaskState:
      type: string
      enum: [pending, running, completed, failed]

    TrackedTask:
      type: object
      required: [id, type, state, createdAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        params:
          type: object
          additionalProperties: true
        state:
          $ref: "#/components/schemas/TaskState"
        error:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time

    EngineStatus:
      type: object
      required: [phase]
      properties:
        phase:
          type: string
        currentTask:
          type: string
        lastTask:
          type: string
        lastTaskResult:
          type: string
        lastTaskError:
          type: string
        activeTasks:
          type: integer
        upgradeHeight:
          type: integer
          format: int64
        upgradeImage:
          type: string
        pendingUpgrades:
          type: integer

    CreateScheduleRequest:
      type: object
      required: [taskType]
      properties:
        taskType:
          type: string
        params:
          type: object
          additionalProperties: true
        cron:
          type: string
          description: Cron expression (mutually exclusive with runAt).
        runAt:
          type: string
          format: date-time
          description: One-shot execution time (mutually exclusive with cron).

    Schedule:
      type: object
      required: [id, taskType, enabled, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskType:
          type: string
        params:
          type: object
          additionalProperties: true
        cron:
          type: string
        runAt:
          type: string
          format: date-time
        enabled:
          type: boolean
        lastRunAt:
          type: string
          format: date-time
        nextRunAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string